[consensus] rename experimental to pipeline
