+++ aptos-move/framework/src/aptos.rs
@@ -73,7 +73,7 @@ impl ReleaseTarget {
                 Some("cached-packages/src/aptos_token_objects_sdk_builder.rs"),
             ),
         ];
-        // Currently we don't have experimental packages only included in particular targets.
+        // Currently we don't have pipeline packages only included in particular targets.
         result
     }
