+++ consensus/src/lib.rs
@@ -22,7 +22,6 @@ mod consensusdb;
 mod dag;
 mod epoch_manager;
 mod error;
-mod experimental;
 mod liveness;
 mod logging;
 mod metrics_safety_rules;
@@ -32,6 +31,7 @@ mod network_tests;
 mod payload_client;
 mod pending_votes;
 pub mod persistent_liveness_storage;
+mod pipeline;
 pub mod quorum_store;
 mod rand;
 mod recovery_manager;
