+++ consensus/src/pipeline/errors.rs
