+++ consensus/src/pipeline/tests/phase_tester.rs
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
-    experimental::{
+    pipeline::{
         buffer_manager::{Receiver, Sender},
         pipeline_phase::{CountedRequest, StatelessPipeline},
     },
