+++ terraform/helm/genesis/values.yaml
@@ -73,7 +73,7 @@ genesis:
     internal_host_suffix: fullnode-lb
   # -- The local path for move modules in the docker image. Defaults to the aptos-framework in the aptoslabs/tools docker image
   moveModulesDir: /aptos-framework/move/modules
-  # -- Options for multicluster mode. This is *experimental only*
+  # -- Options for multicluster mode. This is *pipeline only*
   multicluster:
     enabled: false
     # comma separated list of cluster names
